pyproject.toml
[tool.poetry]
name = "youtube-audio-telegram"
version = "0.1.0"
description = ""
authors = ["Dmitry Barskov <dmitry.barskov@flatstack.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.10"
pytube = "^11.0.2"
environs = "^9.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
